Created new Flask project for Growlin
This Flask project may replace the other (Django) project, if I'm
happy with it. Right now, I'm just going to be experimenting with
Flask for a while to see how it works. I think it may be a better
choice than Growlin for an offline app, beacuse it's more minimal
and so more suitable for packaging as a desktop app.

Currently, the Flask app just has basic login/registration
functionality with Flask_Security and peewee, as well as an empty
Admin site using Flask-Admin. Stay tuned for more updates, though!