Login process restructure!
Changed login process to use simple Flask-Login and Flask-Principal
instead of Flask-Security as they are more customizable and so will
be better for fine-tuning the customized Growlin login.

After all, Flask-Security is more for websites, and Growlin will need
some tailor-made login procedures specific to the library-program. So
it makes more sense to write the whole process from scratch instead of
taking a framework designed for a different use and trying to override
it to suit our needs.