Removed Primary Key attribute from Copy.accession
It's better if Copy has a separate Primary Key, for security
reasons. Otherwise, the accession will be used whenever the Copy
is referenced, meaning program users may have access to it via the
URL or other hidden form fields.

The Accession Number is generally not supposed to be viewed by
non-admin end-users; it is only used indirectly when users enter an
accession number to be checked against the records. In a desktop
application it's all right to use the accession as a primary key and
instruct the app not to display it, but for a browser-based app it's
better not to reveal it in the source either.