Added Group-sorted user list to login page
Also added JS so that when a username is clicked on, name is
automatically filled into the form field, and password field is
shown or form is submitted, as required (depending on whether the
user has a password or not).

Right now, the username/password fields are *also* visible: page
styling is yet to be done.