Edited names and replaced pk fields with unique contraints
Some of the field names like PublicationType.itype and
Location.loc_name have been all renamed to 'name' to make them
easier to remember. Also, a lot of the fields that were being used
as primary keys now have the pk property removed and a unique
constraint set instead. This is to make it easier to implement
field editing and renaming. ForeignKeys which were taking data from
the PK can now be made to point to the unique-constrained field
instead.