Restructured type-specific publication data storage
Type-specific data (eg. book publishers or magazine date) can now
be stored in separate models created specifically for those types,
instead of cluttering up the Publication and Copy models. This extra
data is optional, and the program can function smoothly even without
these models.

New models for holding this data can be created and added to the
`all_pubtypes` and `all_pubcopies sets respectively for publication-
and copy-specific data. TODO: automating the process of adding the
new models to the sets.