Added ReverseGFK's for BaseCopyType and BasePubType
That's just the short version. I also moved around the models a bit
and added a catch-case in the admin registration in case a pubtype
or pubcopy model doesn't have a verbose_name defined.

Also, I reordered the models in the file. And to implement having
a model-specific ReverseGFK for *every* field, I did a lot of
copying from peewee's playhouse.gfk addon, using peewee's internal
`with_metaclass` function which I don't fully understand...anyway,
at least it's working. for now.

To add a new pubtype/pubcopy model, make it subclass BasePubType or
BaseCopyType, and then register it by doing all_pubtypes.add(MyModel)
or all_pubcopies.add(MyModel) as required. This is yet to be implemented
in the admin interface (or anywhere else, for that matter) but will be
soon.