Added "borrow" and "unborrow" functions for the User model
Since models are supposed to handle the data, it makes sense to
handle all these kinds of operations there, instead of doing it in
the views interface. Eventually, the views interface can just try
the operations and then return the errors (such as 'You have not
borrowed that item!').

The way I'm doing it is so that it'll also be possible to program
other user interfaces (eg. CLI) for the same growlin database and
using the same models file, without having to repeat too much of the
logic code.

TODO: Right now the errors are returned as ValueError, IntegrityError,
etc., but perhaps it'll be better to have one BorrowError exception class
that can be caught more easily from the UI code. I'll do this once I figure
out how to write custom exceptions...