Create SQL table definitions file (tables.sql)
This file has currently been populated with sql auto-generated from
the peewee models code. Please note that no changes have been made
to the auto-generated code yet, except for expanding it into many
lines for better readability. Further optimization will be following
soon in a series of commits, including changing the structure to
match the one discussed in the meeting of 2015-08-07....