SQL: Merge multiple publication/copy tables into "item"
This is to fit to the new database structure we decided on in the
meeting of 2015-08-07. Bascially, each item (it can be a book,
magazine, article, or anything else with an accession number) has
its own record in the table. If a particular item type has extra
properties that need to be stored (eg. magazine date), an extra
field for that property can be added to the "item" table, ideally
using varchar2 so that it doesn't take up extra space when not in
use.

To avoid displaying multiple results of the same book, if there
happen to be multiple copies registered in the library, a separate
view will be created which merges the copies into one record,
displaying the number of copies existing and (later) available and/or
borrowed.

@KrishnaRamamoorthy: could you write the code for the materialized
view? Code for counting how many are borrowed and how many are
available, etc. can be added later, but we should at least get the
main view going.