Remove intermediate "Welcome to Growlin" page
Now, when you enter Growlin you go straight to the login page, and
when you sign in you go straight to your user shelf. There's no
intermediary "Welcome to Growlin, click here to sign in" page. It
was never meant to be there anyway, except to indicate that Flask
was working. But it never got removed...till now.