Tweak PastBorrowing model
Now it has a concrete ForeignKey to the Publication (not to the copy,
since you usually won't need to know which particular copy you borrowed)
instead of having just an IntegerField which we know to be the same. I
think this will help write queries more easily.

Also, User now has a `get_past_borrowings()` method, similar to the
`get_current_borrowings()` one. (To return to the queries thing: yes,
it *was* easier!)