Change "More info" icon in shelf/history.htm
Sorry, the other way i tried didn't work...I still have to get the
hang of Materialize!

@abhiniype, why was the style="material-icons return" instead of
just "material-icons"? I thought the "return" was the icon name,
but from what I can figure out the icon gets set from the *content*
of the <i>, not from the style.

Or was it just a mistake?