SQL: Trim currency
TL;DR version: Only name and symbol are required. Growlin is a library
app, not a stock-market app. Storing `conversion_factor` will only
make sense if we store the date as well, since the conversion rate
keeps changing.

So `currency` will only be used to represent a certain currency,
and recorded prices will be how much was paid at the time of the
transaction, without any internal conversion. Anyway,the conversion
rate can always be deduced later if we really need to.