Changed models.py models to use MongoEngine
Class methods and test data generators still need to be updated,
as does some of the query code within growlin.py. Speaking of which,
from now on I'll try to rewrite it so that all the database-specific
code is handled within models.py itself, so the database can be
swapped with minimal changes to the other code.