Add testing script (test.py)
This is what I used instead of typing stuff in the prompt every
time. I haven't done testing before so don't expect this script to
be very well-designed: I just added it "by the way" since it'll
probably come in useful later on (I'll probably update it to use
a proper testing framework, once I figure out how that works...).