Update BaseModel to include a MongoEngine-compatible `objects` attribute
Currently, only `get` is supported. There won't be a 1:1
compatibility with MongoEngine; the plan is to develop for MongoEngine
and then add the relavant functions for the MySQL version as required.

This is mainly a wrapper so that the version of Growlin developed for
MongoEngine works for MySQL as well without changes to any of the
other files.