Restructure and modularize app
Code is now sorted into separate modules instead of being all lumped
into one huge file. Relative imports are added, and the new run.py
script is to be used to run the app. Configuration is now to be done
on-site in the `instance/` folder (not included, except in .gitignore)