Update ItemType model and add get_item_types() method
This is so that other parts of the program know what ItemTypes are
available (in case eg. someone subclasses ItemType and adds more
models). This function could possibly be updated later to dynamically
get the list of item types from eg. plugin import details or something.
Assuming that Growlin begins to support plugins at some point, that is!