Add embedded related model creator for Select2 fields
It sort of works, with a bit of extra config. Also, all the related
models that you are editing have to be enabled with edit_modal=True
(you'll see in the next commit).

The /api/ok view was also added as a dummy so that the AJAX function
recognises a successful insertion