Drop MySQL support and remove peewee implementation
Growlin is going to be MongoDB-only, and will no longer support
MySQL and other RDBMS databases. This is to make the code simpler,
and also due to the fact that the people who wanted MySQL support
in the first place have decided they're not going to use the
application after all.